Split single crate into hyalo-core and hyalo-cli workspace
Reorganize the monolithic crate into a Cargo workspace with two members:
hyalo-core (library) and hyalo-cli (binary). This enforces module
boundaries at the crate level and enables faster incremental builds.
