+++ Cargo.toml
@@ -1,30 +1,7 @@
-[package]
-name = "hyalo"
-version = "0.1.0"
-edition = "2024"
-license = "MIT"
-publish = false
-
-[dependencies]
-anyhow = "1"
-clap = { version = "4.6.0", features = ["derive"] }
-dunce = "1"
-globset = "0.4"
-ignore = "0.4"
-jaq-core = "2.2.1"
-regex = "1"
-jaq-json = { version = "1.1.3", features = ["serde_json"] }
-jaq-std = "2.1.2"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1.0.149"
-serde_yaml_ng = "0.10" # depends on unsafe-libyaml; monitor for pure-Rust alternatives
-tempfile = "3"
-toml = "0.8"
-
-[dev-dependencies]
-assert_cmd = "2"
-criterion = { version = "0.5", features = ["html_reports"] }
-predicates = "3"
+[workspace]
+resolver = "3"
+members = ["crates/*"]
+default-members = ["crates/hyalo-cli"]
 
 [profile.release]
 codegen-units = 1
@@ -35,11 +12,3 @@ strip = true
 [profile.bench]
 codegen-units = 1
 lto = true
-
-[[bench]]
-name = "micro"
-harness = false
-
-[[bench]]
-name = "vault"
-harness = false
