+++ crates/hyalo-cli/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "hyalo-cli"
+version = "0.1.0"
+edition = "2024"
+license = "MIT"
+publish = false
+
+[[bin]]
+name = "hyalo"
+path = "src/main.rs"
+
+[dependencies]
+hyalo-core = { path = "../hyalo-core" }
+anyhow = "1"
+clap = { version = "4.6.0", features = ["derive"] }
+jaq-core = "2.2.1"
+jaq-json = { version = "1.1.3", features = ["serde_json"] }
+jaq-std = "2.1.2"
+regex = "1"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1.0.149"
+serde_yaml_ng = "0.10"
+toml = "0.8"
+
+[dev-dependencies]
+assert_cmd = "2"
+predicates = "3"
+tempfile = "3"
