+++ crates/hyalo-cli/src/commands/summary.rs
@@ -5,12 +5,12 @@ use std::path::Path;
 use std::time::SystemTime;
 
 use crate::commands::{FilesOrOutcome, collect_files};
-use crate::filter::extract_tags;
-use crate::frontmatter::{infer_type, yaml_to_json};
 use crate::output::{CommandOutcome, Format};
-use crate::scanner::{FrontmatterCollector, scan_file_multi};
-use crate::tasks::TaskCounter;
-use crate::types::{
+use hyalo_core::filter::extract_tags;
+use hyalo_core::frontmatter::{infer_type, yaml_to_json};
+use hyalo_core::scanner::{FrontmatterCollector, scan_file_multi};
+use hyalo_core::tasks::TaskCounter;
+use hyalo_core::types::{
     DirectoryCount, FileCounts, PropertySummaryEntry, RecentFile, StatusGroup, TagSummary,
     TagSummaryEntry, TaskCount, VaultSummary,
 };
