+++ crates/hyalo-cli/src/hints.rs
