+++ crates/hyalo-cli/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod commands;
+pub mod config;
+pub mod hints;
+pub mod output;
