+++ crates/hyalo-cli/src/output.rs
