+++ crates/hyalo-cli/tests/e2e_append.rs
