+++ crates/hyalo-cli/tests/e2e_errors.rs
