+++ crates/hyalo-cli/tests/e2e_find.rs
