+++ crates/hyalo-cli/tests/e2e_help.rs
