+++ crates/hyalo-cli/tests/e2e_hints.rs
