+++ crates/hyalo-cli/tests/e2e_properties.rs
