+++ crates/hyalo-cli/tests/e2e_remove.rs
