+++ crates/hyalo-core/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "hyalo-core"
+version = "0.1.0"
+edition = "2024"
+license = "MIT"
+publish = false
+
+[dependencies]
+anyhow = "1"
+dunce = "1"
+globset = "0.4"
+ignore = "0.4"
+regex = "1"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1.0.149"
+serde_yaml_ng = "0.10"
+tempfile = "3"
+
+[dev-dependencies]
+criterion = { version = "0.5", features = ["html_reports"] }
+
+[[bench]]
+name = "micro"
+harness = false
+
+[[bench]]
+name = "vault"
+harness = false
