+++ crates/hyalo-core/src/content_search.rs
