+++ crates/hyalo-core/src/discovery.rs
