+++ crates/hyalo-core/src/heading.rs
