+++ crates/hyalo-core/src/lib.rs
@@ -1,14 +1,10 @@
-pub mod commands;
-pub mod config;
 pub mod content_search;
 pub mod discovery;
 pub mod filter;
 pub mod frontmatter;
 pub mod fs_util;
 pub mod heading;
-pub mod hints;
 pub mod links;
-pub mod output;
 pub mod scanner;
 pub mod tasks;
 pub mod types;
