+++ crates/hyalo-core/src/links.rs
